core: fix autopilot for missions and params
It turns out we were using our instead of the system's autopilot type
(e.g. PX4 or ArduPilot) in the transfer from the mission client and
param client.

This should fix the timeout happening for ArduPilot (as ArduPilot
responds to MISSION_COUNT with MISSION_REQUEST instead of
MISSION_REQUEST_INT).

Signed-off-by: Julian Oes <julian@oes.ch>