third_party: fix illegal instruction on RPi 4
The absl library tries to make use of the crypto libraries on arm64
platforms. However, it turns out that the RPi 4 chip does not have
the crypto extensions. Hence, we need to patch absl to refrain from
using the hardware crypto stuff and fallback to slower software
implementations. This is only required for a random back-off timer at
the moment, so shouldn't matter too much.

Note that the RPi 5 has the crypto extensions but our build is the same
for both.

Signed-off-by: Julian Oes <julian@oes.ch>