core: potentially fix timeout handler
I have seen curious cases where timeouts just didn't seem to trigger as
the they should have. My guess was that the void * pointer to the
location of the shared_ptr in the map was just a bad idea in the first
place. So I changed it to just a uint64_t handle which is easier to
debug and should always be unique. I also changed from a std::map to a
std::vector, however, the jury is out on what is actually faster for our
usage.